Adds support for WebVR mappings
This change moves existing mappings into a WebXR subfolder and renames
them to conform with WebXR naming conventions. It then adds a WebVR
subfolder in which new WebVR mappings are added based on the known
gamepad ids in released browsers. The schema is extended to add an
property indicating a mapping is for WebVR. When this property is true,
additional WebVR-specific properties for targetRay origin and axis
inversion are added to the valid schema.
Tests are added and test tools are modified to look for  mapping files
in both WebVR and WebXR subfolders.