Added a variety of backgrounds for the model viewer
These backgrounds show what the controllers look like in a variety of
lighting conditions, which can help when debugging material issues. The
previous lighting model (with one light directly below and one directly
above) was causing some controllers to appear with a black seam down
the middle, which was causing me to think my normals had an issue when
really it was just the lights. This change resolves that while simply
looking a bit nicer.