Updating package build variants to remove footguns.
Previously if you just ran "npm run build" it would fail because the lerna's bootstrap hasn't been run yet. This change makes the default build run the bootstrap, adds an incrementalBuild that does what build previously did (because the bootstrap can be slow) and changes cleanBuild to run "npm run clean", which includes the bootstrap. This should make it easier for devs to naturally stumble into the right behavior. (Pit of Success!)