Added assets for the Vive Focus
Assets pulled from https://github.com/aframevr/assets/tree/master/controllers/vive/focus-controller

To be honest, this model is significantly more detailed than the platform warrants. I took the time to remove some of the most unnecessary details where I could, but properly decimating the mesh in such a way that it still looks good is going to be far more time consuming than I'm willing to put into it, given the number of users I expect will actually encounter this mesh in the wild. Mostly just pushing to make sure we publish the library with assets for every profile.