Add generic-touchscreen input source profile
It's based on generic-touchpad profile - we could reuse it as-is, but it's
probably better to have a separate profile in this case as touchscreen and
touchpad are conceptually different.

The touchscreen could expose a gamepad object - this should make it easier
for the apps to compute screen space coordinates of the touched point, or
use the values as-is if that's sufficient for their use case.