Further simplified the asset
Button icons and logo text are now baked into the texture rather than being physically modeled. Still could stand to have a bit more geometry reduced (we don't need physical seams all over the device, for example) but I think this is as far as I'm willing to take it.