Updated fetchProfile to allow a defaultProfile to be passed
This is the profile that'll load if none of the profile in the inputSource are matched. Debated about whether or not to pick a default but this feels like the least surprising way forward, as otherwise we might create a lot of false-positive controller "hits" on devices that don't actually match what's loaded at all. Doing it this way allows the developer to be intentional about their desire to fallback and select a profile that matches their needs most closely.