Enable profiles to indicate deprecated profile ids
/fixes #116

For cases where user agents report a non-standard name for some period
of time (which has already been the case with at least one profile in
Chrome) the deprecatedProfileIds let us identify a profile id as being
associated with the assets without implying that it's a registered
profile id.

This PR changes the format of the profilesList.json files to make each
value a dictionary with a `path` key rather than just a string. This
allows us to also report that a particular id is deprecated (so we can
skip it for uses like the profile viewer). I suspect we'll find it
useful in the future for additional metadata as well.