Cache the profile list and profiles when loaded.
I noticed that certain uses of VR devices was causing a lot of network traffic as we fetched the same list and profiles over and over again. This was most common when pulling up the system menu, which would disconnect the controllers from the web app, and then reconnect them when you dismissed the menu. Every time this happened there would be a significant pause before the controller assets would reappear again. Most of that is likely the rendering library re-parsing the glTF files, which I'll address separately, but this seemed like a reasonable point to do a bit of caching as well.